refactor(tools): consolidate 3 per-engine connect-gce tools into cloud-sql-connect-gce
Replaces cloud-sql-{postgres,mysql,mssql}-connect-gce with a single
cloud-sql-connect-gce tool that auto-detects the engine from the
DatabaseVersion returned by the Cloud SQL Admin API. Callers no longer
have to know or pass the engine; the LLM cannot pick the wrong sibling
tool for a given instance.

The new tool lives in internal/tools/cloudsql/cloudsqlconnectgce
(alongside the other cloudsqladmin-backed tools) and:

- Runs the strict engine parse and rejects unknown DatabaseVersion
  values up front.
- Uses the token-scoped GetComputeService, so both the Cloud SQL Admin
  and Compute Engine calls run as the caller when useClientOAuth is
  set on the source.
- Defaults database_name to the engine's conventional default
  (postgres / mysql / master) after detection and always validates the
  resolved value.
- Retains the same RequiredIAMRoles, RequiredAPIs, and result schema
  the three prior tools returned.

Deletes internal/tools/cloudsqlpg/cloudsqlpgconnectgce,
internal/tools/cloudsqlmysql/cloudsqlmysqlconnectgce, and
internal/tools/cloudsqlmssql/cloudsqlmssqlconnectgce; updates
cmd/internal/imports.go accordingly.

Addresses #3740 review: 1-vs-3 tools thread and the auth-scoping
thread.
