+++ cmd/internal/imports.go
@@ -136,6 +136,7 @@ import (
 	_ "github.com/googleapis/mcp-toolbox/internal/tools/cloudloggingadmin/cloudloggingadminquerylogs"
 	_ "github.com/googleapis/mcp-toolbox/internal/tools/cloudmonitoring"
 	_ "github.com/googleapis/mcp-toolbox/internal/tools/cloudsql/cloudsqlcloneinstance"
+	_ "github.com/googleapis/mcp-toolbox/internal/tools/cloudsql/cloudsqlconnectgce"
 	_ "github.com/googleapis/mcp-toolbox/internal/tools/cloudsql/cloudsqlcreatebackup"
 	_ "github.com/googleapis/mcp-toolbox/internal/tools/cloudsql/cloudsqlcreatedatabase"
 	_ "github.com/googleapis/mcp-toolbox/internal/tools/cloudsql/cloudsqlcreateusers"
@@ -146,11 +147,8 @@ import (
 	_ "github.com/googleapis/mcp-toolbox/internal/tools/cloudsql/cloudsqlwaitforoperation"
 	_ "github.com/googleapis/mcp-toolbox/internal/tools/cloudsqladmin/cloudsqladminexecutemany"
 	_ "github.com/googleapis/mcp-toolbox/internal/tools/cloudsqladmin/cloudsqladminsqlmany"
-	_ "github.com/googleapis/mcp-toolbox/internal/tools/cloudsqlmssql/cloudsqlmssqlconnectgce"
 	_ "github.com/googleapis/mcp-toolbox/internal/tools/cloudsqlmssql/cloudsqlmssqlcreateinstance"
-	_ "github.com/googleapis/mcp-toolbox/internal/tools/cloudsqlmysql/cloudsqlmysqlconnectgce"
 	_ "github.com/googleapis/mcp-toolbox/internal/tools/cloudsqlmysql/cloudsqlmysqlcreateinstance"
-	_ "github.com/googleapis/mcp-toolbox/internal/tools/cloudsqlpg/cloudsqlpgconnectgce"
 	_ "github.com/googleapis/mcp-toolbox/internal/tools/cloudsqlpg/cloudsqlpgcreateinstances"
 	_ "github.com/googleapis/mcp-toolbox/internal/tools/cloudsqlpg/cloudsqlpgupgradeprecheck"
 	_ "github.com/googleapis/mcp-toolbox/internal/tools/cloudsqlpg/vectorassistapplyspec"
